Checks for amount and quantity. Rewrite amount with BigDecimal.

+++ src/main/java/com/example/DailySpend/dto/SpendRequestDTO.java
@@ -2,9 +2,11 @@ package com.example.DailySpend.dto;
 
 import jakarta.validation.constraints.NotNull;
 
+import java.math.BigDecimal;
+
 public record SpendRequestDTO(
         @NotNull String name,
-        @NotNull Double amount,
+        @NotNull BigDecimal amount,
         @NotNull Integer quantity,
         @NotNull String category
 ) {
