+++ src/main/java/com/example/DailySpend/dto/SummaryResponseDTO.java
@@ -2,9 +2,10 @@ package com.example.DailySpend.dto;
 
 import com.example.DailySpend.constants.SpendCategory;
 
+import java.math.BigDecimal;
 import java.util.Map;
 
 public record SummaryResponseDTO(
-        Map<SpendCategory, Double> categories,
-        Double total
+        Map<SpendCategory, BigDecimal> categories,
+        BigDecimal total
 ) {}
